first

+++ application/views/template/backend/pages/t_m_d_level_user.php
@@ -1,6 +1,6 @@
 <div class="card">
   <div class="card-header">
-    <h5>Master Divisi</h5>
+    <h5>Master Level User</h5>
   </div>
   <div class="card-block">
     <!-- Menampilkan notif !-->
@@ -12,9 +12,8 @@
       <table id="dom-jqry" class="table table-striped table-bordered nowrap">
         <thead>
           <tr>
-            <th>No</th>
-            <th>Divisi Id</th>
-            <th>Divisi</th>
+            <th>ID</th>
+            <th>Level User</th>
             <th>Action</th>
           </tr>
         </thead>
@@ -22,27 +21,53 @@
           <?php
           foreach ($c_t_m_d_level_user as $key => $value) 
           {
-            echo "<tr>";
-            echo "<td>".($key + 1)."</td>";
-            echo "<td>".$value->LEVEL_USER_ID."</td>";
-            echo "<td>".$value->LEVEL_USER."</td>";
-          
-            echo "<td>";
-             
-            echo "<a href='javascript:void(0);' data-toggle='modal' data-target='#Modal_Edit' class='btn-edit' data-id='".$value->ID."'>";
-              echo "<i class='icon feather icon-edit f-w-600 f-16 m-r-15 text-c-green'></i>";
-            echo "</a>";
+            if($value->MARK_FOR_DELETE == 'f')
+            {
+              echo "<tr>";
+              echo "<td>".$value->ID."</td>";
+              echo "<td>".$value->LEVEL_USER."</td>";
+            
+              echo "<td>";
+               
+              echo "<a href='javascript:void(0);' data-toggle='modal' data-target='#Modal_Edit' class='btn-edit' data-id='".$value->ID."'>";
+                echo "<i class='icon feather icon-edit f-w-600 f-16 m-r-15 text-c-green'></i>";
+              echo "</a>";
 
-            echo "<a href='".site_url('c_t_m_d_level_user/delete/' . $value->ID)."' ";
-            ?>
-            onclick="return confirm('Apakah kamu yakin ingin menghapus data ini?')"
-            <?php
-            echo "> <i class='feather icon-trash-2 f-w-600 f-16 text-c-red'></i></a>";
+              echo "<a href='".site_url('c_t_m_d_level_user/delete/' . $value->ID)."' ";
+              ?>
+              onclick="return confirm('Apakah kamu yakin ingin menghapus data ini?')"
+              <?php
+              echo "> <i class='feather icon-trash-2 f-w-600 f-16 text-c-red'></i></a>";
 
-            echo "</td>";
+              echo "</td>";
 
 
-            echo "</tr>";
+              echo "</tr>";
+            }
+            
+            if($value->MARK_FOR_DELETE == 't')
+            {
+              echo "<tr>";
+              echo "<td><s>".($key + 1)."</s></td>";
+              echo "<td><s>".$value->LEVEL_USER."</s></td>";
+            
+              echo "<td>";
+               
+              
+
+              echo "<a href='".site_url('c_t_m_d_level_user/undo_delete/' . $value->ID)."' ";
+              ?>
+              onclick="return confirm('Apakah kamu yakin ingin mengembalikan data ini?')"
+              <?php
+              echo "> <i class='fa fa-refresh f-w-600 f-16 text-c-red'></i></a>";
+
+              echo ' '.$value->UPDATED_BY;
+              echo "</td>";
+
+
+              echo "</tr>";
+            }
+            
 
           }
           ?>
@@ -70,10 +95,6 @@
         <div class="modal-body">
           <div class="">
 
-            <div class="form-group">
-              <label>Level User ID</label>
-              <input type='text' class='form-control' placeholder='Harus Angka' name='level_user_id'>
-            </div>
 
             <div class="form-group">
               <label>Level User</label>
@@ -112,10 +133,6 @@
 
             <input type="hidden" name="id" value="" class="form-control">
 
-            <div class="form-group">
-              <label>Level User ID</label>
-              <input type='text' class='form-control' placeholder='Harus Angka' name='level_user_id'>
-            </div>
 
             <div class="form-group">
               <label>Level User</label>
@@ -125,6 +142,19 @@
 
           </div>
           <div class="modal-footer">
+            <div class="created_form">
+              Created By : <a name='created_by'></a>
+              <br>
+              Updated By : <a name='updated_by'></a>
+            </div>
+            <style type="text/css">
+              .created_form
+              {
+                float: left;
+                margin right: : 20px;
+                font-size: 10px;
+              }
+            </style>
             <button type="button" class="btn btn-default waves-effect " data-dismiss="modal">Close</button>
             <button type="Submit" class="btn btn-primary waves-effect waves-light ">Save changes</button>
           </div>
@@ -147,13 +177,15 @@
       });
       const {
         ID,
-        LEVEL_USER_ID : level_user_id,
-        LEVEL_USER : level_user
+        LEVEL_USER : level_user,
+        CREATED_BY : created_by,
+        UPDATED_BY : updated_by
       } = User[0];
 
       elModalEdit.querySelector("[name=id]").value = ID;
-      elModalEdit.querySelector("[name=level_user_id]").value = level_user_id;
       elModalEdit.querySelector("[name=level_user]").value = level_user;
+      elModalEdit.querySelector("[name=created_by]").text = created_by;
+      elModalEdit.querySelector("[name=updated_by]").text = updated_by;
 
     })
   })
